emotional_constellation/requirements.txt
streamlit==1.39.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
nltk>=3.8.1
